Add Board interface
